% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{AddMetadataForSeurat}
\alias{AddMetadataForSeurat}
\title{AddMetadataForSeurat function}
\usage{
AddMetadataForSeurat(Obj, FieldName, Values)
}
\arguments{
\item{Obj}{A Seurat object. This is created using the CreateSeuratObject in Seurat.}

\item{FieldName}{A character input. Specifies the name of the field to be added to the metadata}

\item{Values}{A numeric or character variable. Specifies the labels for the cells. The labels should be in the order as the cells in the data set.}
}
\value{
An updated Seurat object containing the metadata with the new field.
}
\description{
This function can be used to add cell metadata information in the Seurat object.
}
\examples{
\dontrun{
Field <- "CellTypes"
CellLabels <- c(rep("B-cells",900),rep("T-cells",900),rep("Monocytes",900))
pbmc3k <- AddMetadataForSeurat(PiccoloList = pbmc3k,FieldName = Field,Values = CellLabels)
}
}
