% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{Bicluster}
\alias{Bicluster}
\title{GenePairsSingleCell function}
\usage{
Bicluster(
  PiccoloList,
  Window = 1,
  MinGenes = 3,
  MinSamples = NULL,
  SampleEnrichment = NULL,
  Out = F
)
}
\arguments{
\item{PiccoloList}{A list object. Piccolo list object obtained after applying the \link[Piccolo]{Normalize} function}

\item{Window}{A numeric (integer) variable. Specifies which window to perform biclustering on. For TuBA, it is set to 1 (default).}

\item{MinGenes}{A numeric variable. Specifies the minimum number of genes that a bicluster should contain. Cannot be less than 3 (default).}

\item{MinSamples}{A numeric variable. Specifies the minimum number of samples that a bicluster must contain. If left NULL (default), no minimum limit is imposed.}

\item{SampleEnrichment}{A numeric variable. Specifies the enrichment level that samples must exhibit in order to belong to a bicluster.}

\item{Out}{A logical variable. If set to T (default is F), will prepare a .csv file containing the bicluster results.}
}
\value{
A list object containing the bicluster results data frame.
}
\description{
This function identifies the gene-pairs with significant overlaps between their percentile sets
}
\examples{
\dontrun{
pbmc3k <- Bicluster(PiccoloList = pbmc3k,SampleEnrichment = 0.05)
}
}
