% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{ComputePC}
\alias{ComputePC}
\title{Compute Principal Components}
\usage{
ComputePC(X, TopFeatures, Barcode, NoOfPC = 50, Out = T)
}
\arguments{
\item{X}{A character variable. Specifies the name of the .csv file containing the standardized values obtained from the \link[Piccolo]{StandardizeMat} function}

\item{TopFeatures}{A character variable. Specifies the name of the .csv file containing the list of genes that were shortlisted as highly variable by the \link[Piccolo]{StandardizeMat} function}

\item{Barcode}{A character variable. Specifies the name of the barcodes file (.tsv format)}

\item{NoOfPC}{A numeric (integer) variable. No of principal components to be retained in the output. Default is 50.}

\item{Out}{A logical variable. Specifies whether to return an output file (.csv) with the normalized values (if set to T), or not (if set to F). Default is T}
}
\value{
A numeric matrix containing the standardized values.
}
\description{
This function will calculate the principal components for the matrix with the standardized values obtained from the \link[Piccolo]{StandardizeMat} function.
}
\examples{
\dontrun{
ComputePC(X = "10X_PBMC3k_logTransformedStandardizedCounts.csv",
TopFeatures = "TopFeatures10X_PBMC3k_matrix.csv",
Barcode = "10X_PBMC3k_barcodes.tsv",Out = T)
}
}
