% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{CreatePiccoloList}
\alias{CreatePiccoloList}
\title{CreatePiccoloList Function}
\usage{
CreatePiccoloList(MTX, Genes, Barcodes, verbose = T)
}
\arguments{
\item{MTX}{A character variable. Specifies the name of the .mtx or .mtx.gz file that contains the counts.}

\item{Genes}{A character variable. Specifies the name of the features (genes) file (.tsv format).}

\item{Barcodes}{A character variable. Specifies the name of the barcodes file (.tsv format).}

\item{verbose}{A logical variable. Specifies whether messages generated while running the function should be displayed (default is T).}
}
\value{
A list containing the counts matrix, the genes list, and the barcodes.
}
\description{
This function creates a list object containing the counts matrix, the features (genes) list, and the barcodes.
}
\examples{
\dontrun{
 pbmc3k <- CreatePiccoloList(X = "10X_PBMC3k_matrix.mtx.gz",
 Gene = "10X_PBMC3k_features.tsv",
 Barcode = "10X_PBMC3k_barcodes.tsv")
 pbmc3k <- CreatePiccoloList(X = "10X_PBMC3k_matrix.mtx.gz",
 Gene = "10X_PBMC3k_features.tsv",
 Barcode = "10X_PBMC3k_barcodes.tsv"
 MinFeaturesPerCell = 100, MT.Perc = 20,RP.Perc = 90) #changing the filtering criteria
}
}
