% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{FilterCells}
\alias{FilterCells}
\title{FilterCells Function}
\usage{
FilterCells(
  PiccoloList,
  MinFeaturesPerCell = 50,
  MT.Perc = 100,
  RP.Perc = 100,
  TotalCountsMADHigh = NULL,
  TotalCountsMADLow = NULL,
  verbose = T
)
}
\arguments{
\item{PiccoloList}{A list object. This should be the list created using the \link[Piccolo]{CreatePiccoloList} function.}

\item{MinFeaturesPerCell}{An integer variable. Specifies the minimum number of genes with non-zero counts within each cell (default is 50).}

\item{MT.Perc}{A numeric variable. Specifies the maximum percentage of total counts within each cell contributed by mitochondrial genes. Cells with MT.Perc greater than specified threshold will be filtered out (default MT.Perc = 100). Can only be used when gene symbols are available in the features file.}

\item{RP.Perc}{A numeric variable. Specifies the maximum percentage of total counts within each cell contributed by ribosomal genes. Cells with RP.Perc greater than specified threshold will be filtered out (default RP.Perc = 100). Can only be used when gene symbols are available in the features file.}

\item{TotalCountsMADHigh}{A numeric variable. Specifies the median absolute deviation (MAD) of the total count above which the cells will be filtered out. Default is NULL, so no filtering based on total count.}

\item{TotalCountsMADLow}{A numeric variable. Specifies the median absolute deviation (MAD) of the total count above which the cells will be filtered out. Default is NULL, so no filtering based on total count.}

\item{verbose}{A logical variable. Specifies whether messages generated while running the function should be displayed (default is T).}
}
\value{
A list containing the filtered counts matrix, the genes list, and the barcodes.
}
\description{
This function is used for filtering cells based on criterias like the minimum number of unique features per cell, the contribution of the mitochondrial and the ribosomal genes to the total counts, as well as identification of outliers based on the total counts.
}
\examples{
\dontrun{
 pbmc3k <- FilterCells(PiccoloList = pbmc3k)
 pbmc3k <- FilterCells(PiccoloList = pbmc3k,
 MinFeaturesPerCell = 100, MT.Perc = 50,
 RP.Perc = 70,TotalCountsMADHigh = 3.5,
 TotalCountsMADLow = 3.5) #changing the filtering criteria
}
}
