% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PiccoloV3.R
\name{GeneSetZScores}
\alias{GeneSetZScores}
\title{Calculate z-scores for gene set}
\usage{
GeneSetZScores(PiccoloList, Genes)
}
\arguments{
\item{PiccoloList}{A list object. Piccolo list object obtained after applying the \link[Piccolo]{Normalize} function.}

\item{Genes}{A character variable. Specifies gene IDs or names. In case of features.tsv or genes.tsv file that contain multiple columns, the input IDs should be present in IDs listed in the first column of the file.}
}
\value{
An updated PiccoloList with a list item containing the combined z-scores for the input gene set.
}
\description{
This function combines and calculates z-scores for gene sets (even works for single genes).
}
\examples{
\dontrun{
GeneSet <- c("ENSG00000123374","ENSG00000135446",
"ENSG00000124762","ENSG00000111276","ENSG00000118007")
pbmc3k <- GeneSetZScores(PiccoloList = pbmc3k,Genes = GeneSet)
GeneSet <- c("CDK2","CDK4","CDKN1A","CDKN1B","STAG1")
pbmc3k <- GeneSetZScores(PiccoloList = pbmc3k,Genes = GeneSet)
}
}
