% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{GetClusterMarkers}
\alias{GetClusterMarkers}
\title{Get marker genes for each cluster}
\usage{
GetClusterMarkers(
  PiccoloList,
  MeanDiffSD = 2.5,
  pValcutoff = NULL,
  FDRcutoff = 0.005,
  MaxSharedClusters = 2,
  Out = F
)
}
\arguments{
\item{PiccoloList}{A list object. Piccolo list object obtained after performing clustering (Leiden or Louvain).}

\item{MeanDiffSD}{A numeric variable. Specifies the minimum number of standard deviations that the difference of mean residuals between cluster cells and non-cluster cells should exhibit in order to be identified as a marker.}

\item{pValcutoff}{A numeric variable. Specifies the minimum raw (uncorrected) p-value that a gene must have to be identified as differentially expressed between the cluster cells and the non-cluster cells. Default is NULL.}

\item{FDRcutoff}{A numeric variable. Specifies the minimum FDR value that a gene must exhibit for it to be considered differentially expressed between the cluster cells and the non-cluster cells. Default is 0.005.}

\item{MaxSharedClusters}{An integer variable. The maximum number of clusters that a given gene may be identified as differentially expressed in. Default is 2.}

\item{Out}{A logical variable. Specifies whether to return an output file (.csv) with the lists of marker genes.}
}
\value{
An updated PiccoloList containing a list with the differential expression result tables containing the marker genes for each cluster.
}
\description{
This function shortlists marker gene sets for each cluster.
}
\examples{
\dontrun{
pbmc3k <- GetClusterMarkers(PiccoloList = pbmc3k,
Out = T)
pbmc3k <- GetClusterMarkers(PiccoloList = pbmc3k,
MaxSharedClusters = 1,Out = T)
}
}
