% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{LabelUMAP}
\alias{LabelUMAP}
\title{UMAP Plot With Cell Labels}
\usage{
LabelUMAP(
  PiccoloList,
  Labels,
  Levels = NULL,
  Alpha = 0.7,
  Size = 1.4,
  BaseSize = 28,
  Title = "Piccolo",
  LegendPosition = "right"
)
}
\arguments{
\item{PiccoloList}{A list object. Piccolo list object obtained after applying the \link[Piccolo]{UMAPcoords} function.}

\item{Labels}{A character variable. Specifies the cell-type/group/batch labels for the cells in the same order that they are listed in the barcodes file.}

\item{Levels}{A character variable. Specifies the order in which the cluster labels should be listed in the UMAP plot.}

\item{Alpha}{A numeric variable. Specifies the transparency of the dots in the UMAP plot. Smaller values lead to greater transparency. Default is 0.7.}

\item{Size}{A numeric variable. Specifies the size of the dots in the UMAP plot. Default is 1.4.}

\item{BaseSize}{A numeric variable. Specifies the base size of the text elements in the UMAP plot. Default is 28.}

\item{Title}{A character variable. Specifies the title of the UMAP plot.}

\item{LegendPosition}{A character variable. Specifies the position in the plot where the legend should be placed. Default is "right".}
}
\value{
A UMAP plot with the cells colored according to the cell labels provided by the user.
}
\description{
This function generates the UMAP plot with the cells labeled by the cluster labels identitied by Leiden clustering.
}
\examples{
\dontrun{
CellLabels <- c("b-cells","b-cells",..,"cd14 monocytes",..,"NK cells",..)
LabelUMAP(PiccoloList = pbmc3k,
Labels = CellLabels,
Levels = c("b-cells","cd14 monocytes","dendritic","NK cells","naive cytotoxic"),
Title = "PBMC3k")
}
}
