% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{LeidenClustering}
\alias{LeidenClustering}
\title{Perform Leiden Clustering}
\usage{
LeidenClustering(
  PiccoloList,
  Resolution = 1,
  ObjectiveFunction = "modularity",
  Weights = NULL,
  Beta = 0.01,
  InitialMembership = NULL,
  N_Iterations = 2,
  VertexWeights = NULL
)
}
\arguments{
\item{PiccoloList}{A list object. Piccolo list object obtained after applying the \link[Piccolo]{KNearestNeighbors} function.}

\item{Resolution}{An numeric variable. An optional parameter that allows the user to adjust the resolution parameter:. Higher resolutions lead to more smaller communities, while lower resolutions lead to fewer larger communities.}

\item{ObjectiveFunction}{A character variable. Specified whether to use the Constant Potts Model (CPM) or modularity. Must be either "CPM" or "modularity".}

\item{Weights}{A positive numeric vector. The weights of the edges. If it is NULL and the input graph has a ‘weight’ edge attribute, then that attribute will be used. If NULL and no such attribute is present, then the edges will have equal weights. Set this to NA if the graph was a ‘weight’ edge attribute, but you don't want to use it for community detection. A larger edge weight means a stronger connection for this function.}

\item{Beta}{A numeric variable. Specifies the parameter affecting the randomness in the Leiden algorithm. This affects only the refinement step of the algorithm.}

\item{InitialMembership}{If provided, the Leiden algorithm will try to improve this provided membership. If no argument is provided, the aglorithm simply starts from the singleton partition.}

\item{N_Iterations}{An integer variable. Specifies    the number of iterations to iterate the Leiden algorithm. Each iteration may improve the partition further.}

\item{VertexWeights}{The vertex weights used in the Leiden algorithm. If this is not provided, it will be automatically determined on the basis of the objective_function. Please see the details of this function how to interpret the vertex weights.}
}
\value{
An updated PiccoloList containing a communities object which contains information about the cluster memberships of the cells.
}
\description{
This function performs Leiden clustering on the graph based on the k nearest neighbors identified using \link[Piccolo]{KNearestNeighbors} function.
}
\examples{
\dontrun{
pbmc3k <- LeidenClustering(PiccoloList = pbmc3k)
pbmc3k <- LeidenClustering(PiccoloList = pbmc3k,
Resolution = 1.5)
}
}
