% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PiccoloV3.R
\name{LouvainClustering}
\alias{LouvainClustering}
\title{Perform Louvain Clustering}
\usage{
LouvainClustering(PiccoloList, Resolution = 1)
}
\arguments{
\item{PiccoloList}{A list object. Piccolo list object obtained after applying the \link[Piccolo]{KNearestNeighbors} function.}

\item{Resolution}{An numeric variable. An optional parameter that allows the user to adjust the resolution parameter of the modularity function that the algorithm uses internally. Lower values typically yield fewer, larger clusters. The original definition of modularity is recovered when the resolution parameter is set to 1.}
}
\value{
An updated PiccoloList containing a communities object which contains information about the cluster memberships of the cells.
}
\description{
This function performs Louvain clustering on the graph based on the k nearest neighbors identified using \link[Piccolo]{KNearestNeighbors} function.
}
\examples{
\dontrun{
pbmc3k <- LouvainClustering(PiccoloList = pbmc3k)
pbmc3k <- LouvainClustering(PiccoloList = pbmc3k,
Resolution = 1.5)
}
}
