% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{LouvainUMAP}
\alias{LouvainUMAP}
\title{UMAP for Louvain Clusters}
\usage{
LouvainUMAP(
  PiccoloList,
  Levels = NULL,
  Alpha = 0.7,
  Size = 1.4,
  BaseSize = 28,
  Title = "Piccolo",
  LegendPosition = "bottom"
)
}
\arguments{
\item{PiccoloList}{A list object. Piccolo list object obtained after applying the \link[Piccolo]{LouvainClustering} function.}

\item{Levels}{A character variable. Specifies the order in which the cluster labels should be listed in the UMAP plot.}

\item{Alpha}{A numeric variable. Specifies the transparency of the dots in the UMAP plot. Smaller values lead to greater transparency. Default is 0.7.}

\item{Size}{A numeric variable. Specifies the size of the dots in the UMAP plot. Default is 1.4.}

\item{BaseSize}{A numeric variable. Specifies the base size of the text elements in the UMAP plot. Default is 28.}

\item{Title}{A character variable. Specifies the title of the UMAP plot.}

\item{LegendPosition}{A character variable. Specifies the position in the plot where the legend should be placed. Default is "bottom".}
}
\value{
A UMAP plot with the cells colored according to the clusters they belong to as identified using the Louvain algorithm.
}
\description{
This function generates the UMAP plot with the cells labeled by the cluster labels identitied by Louvain clustering.
}
\examples{
\dontrun{
LouvainUMAP(PiccoloList = pbmc3k)
pbmc3k <- LeidenClustering(PiccoloList = pbmc3k,
Resolution = 1.5)
}
}
