% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{MaxMinNormMat}
\alias{MaxMinNormMat}
\title{Max-Min Normalization Function}
\usage{
MaxMinNormMat(X, TopFeatures, Barcode, Out = F)
}
\arguments{
\item{X}{A character variable. Specifies the name of the .csv file containing the standardized values obtained from the \link[Piccolo]{StandardizeMat}}

\item{TopFeatures}{A character variable. Specifies the name of the .csv file containing the list of genes that were shortlisted as highly variable by the}

\item{Barcode}{A character variable. Specifies the name of the barcodes file (.tsv format)}

\item{Out}{A logical variable. Specifies whether to return an output file (.csv) with the normalized values (if set to T), or not (if set to F). Default is F}
}
\value{
A numeric matrix containing the standardized values.
}
\description{
This function will normalize the standardized values to normalized values in the range 0-1
}
\examples{
\dontrun{
MaxMinNormMat(X = "10X_PBMC3k_logTransformedStandardizedCounts.csv",
TopFeatures = "TopFeatures10X_PBMC3k_matrix.csv",
Barcode = "10X_PBMC3k_barcodes.tsv",Out = T)
}
}
