% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{NormalizeForSeurat}
\alias{NormalizeForSeurat}
\title{NormalizeForSeurat function}
\usage{
NormalizeForSeurat(
  Obj,
  Transform = "log",
  SizeFactors = NULL,
  verbose = T,
  Out = F
)
}
\arguments{
\item{Obj}{A Seurat object. This should be the list created using the CreateSeuratObject function of Seurat.}

\item{Transform}{A character variable. Specifies the variance stabilizing transformation that will be applied to the counts. The default is the log transform (Transform = "log"). Other options include the Sqrt transform (Transform = "sqrt") and the Box-Cox power law transform (Transform = "bc").}

\item{SizeFactors}{A numeric variable. Can be used to specify size factors per cell obtained from another method. Should be specified in the same order as the cells listed in the barcodes file..}

\item{verbose}{A logical variable. Specifies whether messages generated while running the function should be displayed (default is T).}

\item{Out}{A logical variable. Specifies whether to return output file (.csv) with the z-scores (if set to T), or not (if set to F). Default is FALSE.}
}
\value{
An updated Seurat object containing the matrix with the residuals (z-scores) obtained from the counts matrix in the scale.data slot in the SCT assay.
}
\description{
This function performs normalization for the counts of the genes shortlisted using the \link[Piccolo]{SelectFeaturesForSeurat} function.
}
\examples{
\dontrun{
pbmc3kSeurat <- NormalizeForSeurat(Obj = pbmc3kSeurat)
pbmc3kSeurat <- NormalizeForSeurat(Obj = pbmc3kSeurat,
Transform = "bc")
}
}
