% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{PerformDiffExp}
\alias{PerformDiffExp}
\title{Identify differentially expressed genes between 2 groups of cells}
\usage{
PerformDiffExp(
  PiccoloList,
  Group1,
  Group2,
  Transform = "log",
  Method = "t.test",
  Out = F
)
}
\arguments{
\item{PiccoloList}{A list object. Piccolo list object obtained after applying the \link[Piccolo]{Normalize} function.}

\item{Group1}{A numeric (integers) vector. Specifies the serial numbers of cells in group 1 (serial numbers based on the order of barcodes).}

\item{Group2}{A numeric (integers) vector. Specifies the serial numbers of cells in group 2 (serial numbers based on the order of barcodes).}

\item{Transform}{A character variable. Should be the same as the one used during normalization. Else, the default of "log" will be used.}

\item{Method}{A character variable. Specifies the method used for testing differences between the expression levels in the 2 groups. Currently, two tests are available: the Student's t-test ("t.test) and the Wilcoxon rank-sum test ("wilcoxon").}

\item{Out}{A logical variable. Specifies whether to return an output file (.csv) with the differential expression result (if set to T). Default is F.}
}
\value{
An updated PiccoloList containing a data frame with the results of the differential expression analysis.
}
\description{
This function performs differential expression analysis between 2 groups of cells provided by the user.
}
\examples{
\dontrun{
Group1.vec <- 1:200
Group2.vec <- 301:500
pbmc3k <- PerformDiffExp(PiccoloList = pbmc3k,
Group1 = Group1.vec,
Group2 = Group2.vec,
Out = T)
}
}
