% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{PerformDiffExpClusterwise}
\alias{PerformDiffExpClusterwise}
\title{Differential expression analysis between cells in each cluster vs the rest.}
\usage{
PerformDiffExpClusterwise(
  PiccoloList,
  Transform = "log",
  Method = "t.test",
  Out = F,
  verbose = T
)
}
\arguments{
\item{PiccoloList}{A list object. Piccolo list object obtained after applying the \link[Piccolo]{LeidenClustering} or \link[Piccolo]{LouvainClustering} functions.}

\item{Transform}{A character variable. Should be the same as the one used during normalization. Else, the default of "log" will be used.}

\item{Method}{A character variable. Specifies the method used for testing differences between the expression levels in the 2 groups. Currently, two tests are available: the default Student's t-test ("t.test"), and the Wilcoxon rank-sum test ("wilcoxon").}

\item{Out}{A logical variable. Specifies whether to return an output files (.csv) with the differential expression results (if set to T) or not (set to F). Default is F.}

\item{verbose}{A logical variable. Specifies whether messages generated while running the function should be displayed (default is T).}
}
\value{
An updated PiccoloList containing a data frame with the results of the differential expression analysis.
}
\description{
This function performs differential expression analysis between cells belonging to each cluster (as identified by Leiden or UMAP clustering) and the rest of the cells.
}
\examples{
\dontrun{
pbmc3k <- PerformDiffExpClusterwise(PiccoloList = pbmc3k,
Out = T)
pbmc3k <- PerformDiffExpClusterwise(PiccoloList = pbmc3k,
Method = "wilcoxon", Out = T)
}
}
