% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo.R
\name{pics.coloc.lite}
\alias{pics.coloc.lite}
\title{pics.coloc.lite : Test for colocalization of two PICS sets}
\usage{
pics.coloc.lite(data1, data2, pics1 = "PICS_probability",
  pics2 = "PICS_probability", rsid1 = "Linked_SNP", rsid2 = "Linked_SNP",
  rounded = 6, priorc1 = 1e-04, priorc2 = 1e-04, priorc12 = 1e-05)
}
\arguments{
\item{data1}{PICS sets, from pics.read or pics.download}

\item{data2}{PICS sets, from pics.read or pics.download}
}
\value{
H3 & H4 posterior colocalization info. For concise dataframe returing *only* H3 & H4 posteriors use pics.coloc.lite()

H3 & H4
}
\description{
Same arguements as \code{\link{pics.coloc}}, return object is the only difference.
}
\examples{
myColoc  <- pics.coloc(gwas.credSet, eQTL.credSet)

myColocs <- data \%>\% group_by() \%>\% do(pics.coloc.lite(.$gwas_pics \%>\% .[[1]], .$eqtl_pics \%>\% .[[1]]))

}
