% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo.R
\name{pics.coloc}
\alias{pics.coloc}
\title{pics.coloc : Test for colocalization of two PICS sets
Adapted from [Toby Johnson's code](https://github.com/tobyjohnson/gtx/blob/master/R/abf.R "Toby's coloc in R")
Which was adapted from [Giambartolomei et al. 2014](https://www.ncbi.nlm.nih.gov/pubmed/24830394 "Giambartolomei et al. 2014") colocalization method}
\usage{
pics.coloc(data1, data2, pics1 = "PICS_probability",
  pics2 = "PICS_probability", rsid1 = "Linked_SNP", rsid2 = "Linked_SNP",
  rounded = 6, priorc1 = 1e-04, priorc2 = 1e-04, priorc12 = 1e-05)
}
\arguments{
\item{data1}{PICS sets, from pics.read or pics.download}

\item{data2}{PICS sets, from pics.read or pics.download}

\item{pics1}{Column name to pull PICS prob from data1 if data not from pics.* Default = "PICS_probability"}

\item{pics2}{Column name to pull PICS prob from data2 if data not from pics.* Default = "PICS_probability"}

\item{rsid1}{Column name to pull rsid from if data not from pics.*            Default = "Linked_SNP"}

\item{rsid2}{Column name to pull rsid from if data not from pics.*            Default = "Linked_SNP"}

\item{rounded}{[#] Decimal points to round posteriors to}

\item{priorc1}{Prior probability for colocalization with siganl for data1   Default = 1e-4}

\item{priorc2}{Prior probability for colocalization with siganl for data2   Default = 1e-4}

\item{priorc12}{Prior probability for colocalization of both signals.        Default = 1e-5}
}
\value{
Full colocalization info. For concise dataframe returing *only* H3 & H4 posteriors use pics.coloc.lite()
}
\description{
pics.coloc : Test for colocalization of two PICS sets
Adapted from [Toby Johnson's code](https://github.com/tobyjohnson/gtx/blob/master/R/abf.R "Toby's coloc in R")
Which was adapted from [Giambartolomei et al. 2014](https://www.ncbi.nlm.nih.gov/pubmed/24830394 "Giambartolomei et al. 2014") colocalization method
}
\examples{
myColoc <- pics.coloc(gwas.credSet, eQTL.credSet)

}
