% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo.R
\name{pics.download}
\alias{pics.download}
\title{pics.download : Query the PICs website and get a PICs credible set for a SNP of interest}
\usage{
pics.download(rsid, pvalue, ancestry = "EUR", output = NA,
  override = FALSE)
}
\arguments{
\item{rsid}{valid rsID. e.g. rs6795744}

\item{pvalue}{p-value for the rsid. Either scientific or -log10 base. e.g. "1E-20", "1e-20", or "20"}

\item{ancestry}{= [EUR, ASN, AFR] Default = EUR}

\item{output}{Specify path and name for the download to be saved to. Default = temporary file.}
}
\value{
PICs credible set
}
\description{
pics.download : Query the PICs website and get a PICs credible set for a SNP of interest
}
\examples{
myCredSet <- pics.download(rsid="rs6795744", pvalue = "20")

}
