% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo.R
\name{pics.read}
\alias{pics.read}
\title{pics.read : read a PICs file}
\usage{
pics.read(x)
}
\arguments{
\item{filepath}{String with the filepath to the data to read.}
}
\value{
PICs dataframe
}
\description{
Read into memory PICs from pics.download(output="X") or manually saved text from the website (copy+paste to a text file, not suggested)
}
\examples{
myCredSet <- pics.read("/full/path/to/PICs_credSet.txt")

}
