% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{PrepareCountsForSeurat}
\alias{PrepareCountsForSeurat}
\title{PrepareCountsForSeurat function}
\usage{
PrepareCountsForSeurat(PiccoloList)
}
\arguments{
\item{PiccoloList}{A list object. This should be the list created using the \link[Piccolo]{CreatePiccoloList} function and optionally processed through the \link[Piccolo]{FilterCells}}
}
\value{
An updated PiccoloList object containing the counts matrix with the gene names and the barcodes.
}
\description{
This function adds gene names and barcodes to the counts matrix in PiccoloList in order to use it as input for the CreateSeuratObject function in Seurat.
}
\examples{
\dontrun{
pbmc3k <- PrepareCountsForSeurat(PiccoloList = pbmc3k)
}
}
