% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{SelectFeaturesForSeurat}
\alias{SelectFeaturesForSeurat}
\title{SelectFeaturesForSeurat function}
\usage{
SelectFeaturesForSeurat(
  Obj,
  NoOfHVG = NULL,
  MinPercNonZeroCells = 0.5,
  Reference = NULL,
  Out = F,
  verbose = T
)
}
\arguments{
\item{Obj}{A Seurat object. This should be the list created using the CreateSeuratObject function in Seurat}

\item{NoOfHVG}{A numeric (integer) variable. The number of variable features to be shortlisted. If left NULL (default), will shortlist based on the threshold of ReferenceLevel.}

\item{MinPercNonZeroCells}{A numeric variable. Specifies the minimum percentage of cells that must have non-zero counts for each gene in the data set. Default is 0.5 (\%).}

\item{Reference}{A numeric variable (value should be greater than 0 but less than 1). Specifies the reference level against which features are identified as variable. Default is the 10th quantile for each bin (Reference = 0.1).}

\item{Out}{A logical variable. Specifies whether to return output files (.csv) with the HVGs and the stable genes (if set to T), or not (if set to F). Default is FALSE.}

\item{verbose}{A logical variable. Specifies whether messages generated while running the function should be displayed (default is T).}
}
\value{
An updated Seurat object containing the information of the shortlisted variable genes in the misc list entities of the assays present in the Seurat object.
}
\description{
This function performs feature selection by identifying highly variable genes and stable genes using a Seurat object.
}
\examples{
\dontrun{
pbmc3kSeurat <- SelectFeaturesForSeurat(Obj = pbmc3kSeurat)
pbmc3kSeurat <- SelectFeaturesForSeurat(Obj = pbmc3kSeurat,
NoOfHVG = 3000, Out = T)
}
}
