% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PiccoloV3.R
\name{SelectStableFeaturesAllCells}
\alias{SelectStableFeaturesAllCells}
\title{SelectStableFeaturesAllCells function}
\usage{
SelectStableFeaturesAllCells(
  PiccoloList,
  Reference = NULL,
  MinPercNonZeroCells = 0.5,
  verbose = F
)
}
\arguments{
\item{PiccoloList}{A list object. This should be the list created using the \link[Piccolo]{CreatePiccoloList} function and processed through the \link[Piccolo]{FilterCells}.}

\item{Reference}{A numeric variable (value should be greater than 0 but less than 1). Specifies the reference level against which features are identified as variable. Default is the 10th quantile for each bin (Reference = 0.1).}

\item{MinPercNonZeroCells}{A numeric variable. Specifies the minimum percentage of cells that must have non-zero counts for each gene in the data set. Default is 0.5 (\%).}

\item{verbose}{A logical variable. Specifies whether messages generated while running the function should be displayed (default is T).}
}
\value{
An updated PiccoloList object containing the data frame with the stable genes.
}
\description{
This function performs feature selection by identifying stable genes across all cells.
}
\examples{
\dontrun{
pbmc3k <- SelectStableFeaturesAllCells(PiccoloList = pbmc3k)
pbmc3k <- SelectStableFeaturesAllCells(PiccoloList = pbmc3k, verbose = T)
}
}
