% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{StandardizeMat}
\alias{StandardizeMat}
\title{Standardization Function}
\usage{
StandardizeMat(
  X,
  Gene,
  Barcode,
  VarFeatures = NULL,
  Transform = "log",
  Batch = NULL,
  ReferenceLevel = NULL,
  MinPercNonZero = 1,
  Out = F
)
}
\arguments{
\item{X}{A character variable. Specifies the name of the .mtx or .mtx.gz file that contains the counts.}

\item{Gene}{A character variable. Specifies the name of the features (genes) file (.tsv format)}

\item{Barcode}{A character variable. Specifies the name of the barcodes file (.tsv format)}

\item{VarFeatures}{A numeric (integer) variable. The number of variable features to be shortlisted.}

\item{Transform}{A character variable. Specifies the non-linear transformation that will be applied to the counts. Currently, we offer the log transform (default) and Yeo-Johnson transform. These can be specified as "log" and "yj", respectively. The default is "log".}

\item{Batch}{An optional character vector. Specifies the batch labels for the cells. The order of the batch labels should match the order of the cells in the counts (or barcodes) file.}

\item{ReferenceLevel}{A numeric variable (value should be greater than 0 but less than 1). Specifies the reference level against which features are identified as variable. Default is the median (ReferenceLevel = 0.5).}

\item{MinPercNonZero}{A numeric variable. Specifies the minimum percentage of cells that must have non-zero counts for each gene in the data set. Default is 1 (\%).}

\item{Out}{A logical variable. Specifies whether to return an output file (.csv) with the standardized values (if set to T), or not (if set to F). Default is F}
}
\value{
A numeric matrix containing the standardized values.
}
\description{
This function performs feature selection and prepares standardized counts matrix for the shortlisted features
}
\examples{
\dontrun{
 StandardizeMat(X = "10X_PBMC3k_matrix.mtx.gz",
 Gene = "10X_PBMC3k_features.tsv",
 BinaryMatrix = "10X_PBMC3k_barcodes.tsv",Out = T)
 StandardizeMat(X = "10X_PBMC3k_matrix.mtx.gz",
 Gene = "10X_PBMC3k_features.tsv",
 BinaryMatrix = "10X_PBMC3k_barcodes.tsv",
 ReferenceLevel = 0.3,MinPercNonZero = 0.5,Out = T)
}
}
