% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PiccoloV3.R
\name{tSNEcoords}
\alias{tSNEcoords}
\title{tSNE coordinates}
\usage{
tSNEcoords(PiccoloList, NoOfPC = NULL, Out = F)
}
\arguments{
\item{PiccoloList}{A list object. Piccolo list object obtained after applying the \link[Piccolo]{Normalize} and the \link[Piccolo]{ComputePC} functions.}

\item{NoOfPC}{An integer variable. Specifies the number of PCs to use for the UMAP. The default is NULL, which corresponds to the use of all the shortlisted principal components.}

\item{Out}{A logical variable. Specifies whether to return an output file (.csv) with the UJMAP coordinates (if set to T), or not (if set to F). Default is F.}
}
\value{
A data frame containing the coordinates of the cells in the first 2 UMAP dimensions.
}
\description{
This function generates the 2-dimensional coordinates for the cells using tSNE.
}
\examples{
\dontrun{
pbmc3k <- tSNEcoords(PiccoloList = pbmc3k,
Out = T)
}
}
