% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PiccoloV3.R
\name{UMAPZScores}
\alias{UMAPZScores}
\title{UMAP plot with z-scores for gene sets}
\usage{
UMAPZScores(
  PiccoloList,
  Name,
  xLabel = T,
  yLabel = T,
  BaseSize = 28,
  UpperLowerSDThreshold = 2.5
)
}
\arguments{
\item{PiccoloList}{A list object. Piccolo list object obtained after applying the \link[Piccolo]{GeneSetZScores} function.}

\item{Name}{A character variable. Specifies the name of the gene or gene set that you would like to display as the title.}

\item{xLabel}{A logical variable. Specifies whether the x-axis label should be displayed (T) or not (F). Default is T.}

\item{yLabel}{A logical variable. Specifies whether the y-axis label should be displayed (T) or not (F). Default is T.}

\item{BaseSize}{A numeric variable. Specifies the base size of the text elements in the UMAP plot. Default is 28.}

\item{UpperLowerSDThreshold}{A numeric variable. Specifies the number of standard deviations (SD) above or below which the z-scores should be capped to the value at the specified SD. Default is 2.5.}
}
\value{
A UMAP plot with the cells colored according to the z-scores based on input gene sets specified in \link[Piccolo]{GeneSetZScores}.
}
\description{
This function generates the UMAP plot showing the z-scores for individual genes or gene sets.
}
\examples{
\dontrun{
UMAPZScores(PiccoloList = pbmc3k,Name = "Cell Cycle")
UMAPZScores(PiccoloList = pbmc3k,Name = "Cell Cycle",
UpperLowerSDThreshold = 3.5)
}
}
