% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{ComputePC}
\alias{ComputePC}
\title{Compute Principal Components}
\usage{
ComputePC(PiccoloList, NoOfPC = 50, Out = T)
}
\arguments{
\item{PiccoloList}{A list object. Piccolo list object obtained after applying the \link[Piccolo]{Normalize} function}

\item{NoOfPC}{A numeric (integer) variable. No of principal components to be retained in the output. Default is 50.}

\item{Out}{A logical variable. Specifies whether to return an output file (.csv) with the normalized values (if set to T), or not (if set to F). Default is T}
}
\value{
A numeric matrix containing the standardized values.
}
\description{
This function will calculate the principal components for the matrix with the standardized values obtained from the \link[Piccolo]{Normalize} function.
}
\examples{
\dontrun{
PiccList <- ComputePC(PiccoloList = PiccList)
PiccList <- ComputePC(PiccoloList = PiccList,NoOfPC = 20,Out = T)
}
}
