% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{ComputePC}
\alias{ComputePC}
\title{ComputePC function}
\usage{
ComputePC(PiccoloList, NoOfPC = 50, Out = F, Center = F, Scale = F)
}
\arguments{
\item{PiccoloList}{A PiccoloList object containing NormCounts data frame in it.}

\item{NoOfPC}{An integer variable. Specifies the number of principal components (PCs) to shortlist. Default is 50.}

\item{Out}{A logical variable. Specifies whether to return output file (.csv) with the coordinates for each cell (if set to T), or not (if set to F). Default is FALSE.}

\item{Center}{A logical variable. Specifies whether to center the rows of the residuals matrix for PCA (default is F).}

\item{Scale}{A logical variable. Specifies whether to scale the rows of the residuals matrix for PCA (default is F).}
}
\value{
An updated PiccoloList containing a list with the output of PCA.
}
\description{
This function performs principal components analysis (PCA) using the residuals matrix obtained with the \link[Piccolo]{Normalize} function.
}
\examples{
\dontrun{
pbmc3k <- ComputePC(PiccoloList = pbmc3k)
pbmc3k <- ComputePC(PiccoloList = pbmc3k, NoOfPC = 20)
}
}
