% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{CreatePiccoloList}
\alias{CreatePiccoloList}
\title{CreatePiccoloList Function}
\usage{
CreatePiccoloList(
  X,
  Gene,
  Barcode,
  MinFeaturesPerCell = 200,
  MT.Perc = 10,
  RP.Perc = 70
)
}
\arguments{
\item{X}{A character variable. Specifies the name of the .mtx or .mtx.gz file that contains the counts.}

\item{Gene}{A character variable. Specifies the name of the features (genes) file (.tsv format)}

\item{Barcode}{A character variable. Specifies the name of the barcodes file (.tsv format)}

\item{MinFeaturesPerCell}{A numeric variable. The minimum number of features with non-zero counts that any given cell is expected to contain. Cells with fewer than the specified number will be filtered out (Default is 200).}

\item{MT.Perc}{A numeric variable. The percentage of total count in any given cell attributed to counts of mitochondrial genes. Cells with MT.Perc greater than the specified number will be filtered out (Default is 10).}

\item{RP.Perc}{A numeric variable. The percentage of total count in any given cell attributed to counts of ribosomal genes. Cells with RP.Perc greater than the specified number will be filtered out (Default is 70).}
}
\value{
A list containing the counts matrix, the genes list, and the barcodes.
}
\description{
This function creates a list object containing the counts matrix, the features (genes) list, and the barcodes
}
\examples{
\dontrun{
 pbmc3k <- CreatePiccoloList(X = "10X_PBMC3k_matrix.mtx.gz",
 Gene = "10X_PBMC3k_features.tsv",
 Barcode = "10X_PBMC3k_barcodes.tsv")
 pbmc3k <- CreatePiccoloList(X = "10X_PBMC3k_matrix.mtx.gz",
 Gene = "10X_PBMC3k_features.tsv",
 Barcode = "10X_PBMC3k_barcodes.tsv"
 MinFeaturesPerCell = 100, MT.Perc = 20,RP.Perc = 90) #changing the filtering criteria
}
}
