% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{DEfeatures}
\alias{DEfeatures}
\title{Identify differentially expressed between 2 groups of cells}
\usage{
DEfeatures(PiccoloList, Group1, Group2, Out = F)
}
\arguments{
\item{PiccoloList}{A list object. Piccolo list object obtained after applying the \link[Piccolo]{Normalize} function}

\item{Group1}{A character vector. Specifies the barcodes of cells in group 1}

\item{Group2}{A character vector. Specifies the barcodes of cells in group 2}

\item{Out}{A logical variable. Specifies whether to return an output file (.csv) with the differential expression result (if set to T), or not (if set to F). Default is F}
}
\value{
A data frame containing the gene IDs, the log2 fold change (FC) of normalized values between group 1 and group 2 (positive FC indicates higher expression in group 1), the p-values from the Mann-Whitney test, and the adjusted p-values (p-adj) after Benjamini-Hochberg correction
}
\description{
This function performs differential expression analysis (using the Mann-Whitney test) between 2 groups of cells provided by the user
}
\examples{
\dontrun{
DE.Genes.df <- DEfeatures(PiccoloList = PiccList,
Group1 = c("Barcode1","Barcode23","Barcode47",..),
Group2 = c("Barcode3,"Barcode7, "Barcode11",..)
Out = T)
}
}
