% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{DEfeatures}
\alias{DEfeatures}
\title{Identify differentially expressed between 2 groups of cells}
\usage{
DEfeatures(X, TopFeatures, Barcode, Group1, Group2, Out = F)
}
\arguments{
\item{X}{A character variable. Specifies the name of the .csv file containing the standardized values obtained from the \link[Piccolo]{StandardizeMat} function}

\item{TopFeatures}{A character variable. Specifies the name of the .csv file containing the list of genes that were shortlisted as highly variable by the \link[Piccolo]{StandardizeMat} function}

\item{Barcode}{A character variable. Specifies the name of the barcodes file (.tsv format)}

\item{Group1}{A character vector. Specifies the barcodes of cells in group 1}

\item{Group2}{A character vector. Specifies the barcodes of cells in group 2}

\item{Out}{A logical variable. Specifies whether to return an output file (.csv) with the differential expression result (if set to T), or not (if set to F). Default is F}
}
\value{
A data frame containing the gene IDs, the log2 fold change (FC) of normalized values between group 1 and group 2 (positive FC indicates higher expression in group 1), the p-values from the Mann-Whitney test, and the adjusted p-values (p-adj) after Benjamini-Hochberg correction
}
\description{
This function performs differential expression analysis (using the Mann-Whitney test) between 2 groups of cells provided by the user
}
\examples{
\dontrun{
DEfeatures(X = "10X_PBMC3k_logTransformedStandardizedCounts.csv",
TopFeatures = "TopFeatures10X_PBMC3k_matrix.csv",
Barcode = "10X_PBMC3k_barcodes.tsv",
Group1 = c("Barcode1","Barcode23","Barcode47",..),
Group2 = c("Barcode3,"Barcode7, "Barcode11",..)
Out = T)
}
}
