% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{LabelUMAP}
\alias{LabelUMAP}
\title{Label cells on UMAP}
\usage{
LabelUMAP(PiccoloList, Labels, Levels = NULL, Alpha = 0.7, Size = 0.9)
}
\arguments{
\item{PiccoloList}{A list object. Piccolo list object obtained after applying the \link[Piccolo]{Normalize}, the \link[Piccolo]{ComputePC}, and the \link[Piccolo]{UMAPcoords} functions}

\item{Labels}{A character vector. Should contain the character labels for cells in the same order as the cells in the counts matrix}

\item{Levels}{An optional character vector. Should specify the unique labels in the order in which the labels of the cells should be presented}

\item{Alpha}{A numeric variable (strictly greater than 0). Specified the transparency of the dots on the UMAP plot. Default Alpha = 0.7}

\item{Size}{A numeric variable (strictly greater than 0). Specified the size of the dots on the UMAP plot. Default Size = 0.9}
}
\value{
A ggplot2 object
}
\description{
This function can be used to label (color) cells on the UMAP plot
}
\examples{
\dontrun{
p <- LabelUMAP(PiccoloList = pbmc3k,
Labels = c("b-cells","b-cells",..,"cd14 monocytes",..,"NK cells",..),
Levels = c("b-cells","cd14 monocytes","dendritic","NK cells","naive cytotoxic"))
}
}
