% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{LabelUMAP}
\alias{LabelUMAP}
\title{Label cells on UMAP}
\usage{
LabelUMAP(X, Labels, Levels = NULL, Alpha = 0.7, Size = 0.9)
}
\arguments{
\item{X}{A character variable. Specifies the name of the .csv file containing the UMAP coordinates obtained from the \link[Piccolo]{UMAPCoords} function}

\item{Labels}{A character vector. Should contain the character labels for cells in the same order as the cells in the counts matrix}

\item{Levels}{An optional character vector. Should specify the unique labels in the order in which the labels of the cells should be presented}

\item{Alpha}{A numeric variable (strictly greater than 0). Specified the transparency of the dots on the UMAP plot. Default Alpha = 0.7}

\item{Size}{A numeric variable (strictly greater than 0). Specified the size of the dots on the UMAP plot. Default Size = 0.9}
}
\value{
A ggplot2 object
}
\description{
This function can be used to label (color) cells on the UMAP plot
}
\examples{
\dontrun{
LabelUMAP(X = "10X_PBMC3k_logTransformedStandardizedCounts_Top50PrinComp_UMAPCoords.csv",
Labels = c("b-cells","b-cells",..,"cd14 monocytes",..,"NK cells",..),
Levels = c("b-cells","cd14 monocytes","dendritic","NK cells","naive cytotoxic"))
}
}
