% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PiccoloV3.R
\name{Normalize}
\alias{Normalize}
\title{Normalize function}
\usage{
Normalize(
  PiccoloList,
  Transform = "log",
  SizeFactors = NULL,
  verbose = T,
  Out = F
)
}
\arguments{
\item{PiccoloList}{A list object. This should be the list created using the \link[Piccolo]{CreatePiccoloList} function and processed through the \link[Piccolo]{FilterCells} and the \link[Piccolo]{SelectFeatures} functions.}

\item{Transform}{A character variable. Specifies the variance stabilizing transformation that will be applied to the counts. The default is the log transform (Transform = "log"). Other options include the Sqrt transform (Transform = "sqrt") and the Box-Cox power law transform (Transform = "bc").}

\item{SizeFactors}{A numeric variable. Can be used to specify size factors per cell obtained from another method. Should be specified in the same order as the cells listed in the barcodes file..}

\item{verbose}{A logical variable. Specifies whether messages generated while running the function should be displayed (default is T).}

\item{Out}{A logical variable. Specifies whether to return output file (.csv) with the z-scores (if set to T), or not (if set to F). Default is FALSE.}
}
\value{
An updated PiccoloList object containing the matrix with the residuals (z-scores) obtained from the counts matrix.
}
\description{
This function performs normalization for the counts of the genes shortlisted using the \link[Piccolo]{SelectFeatures} function.
}
\examples{
\dontrun{
pbmc3k <- Normalize(PiccoloList = pbmc3k)
pbmc3k <- Normalize(PiccoloList = pbmc3k, Transform = "bc")
}
}
