% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{UMAPCoords}
\alias{UMAPCoords}
\title{UMAP coordinates}
\usage{
UMAPCoords(X, Out = T)
}
\arguments{
\item{X}{A character variable. Specifies the name of the .csv file containing the principal components obtained using the \link[Piccolo]{ComputePC} function}

\item{Out}{A logical variable. Specifies whether to return an output file (.csv) with the normalized values (if set to T), or not (if set to F). Default is T}
}
\value{
A data frame containing the coordinates of the cells for the first 3 UMAP dimensions.
}
\description{
This function will run UMAP for the matrix with the principal components
}
\examples{
\dontrun{
UMAPCoords(X = "10X_PBMC3k_logTransformedStandardizedCounts_Top50PrinComp.csv",
Out = T)
}
}
