% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Piccolo.R
\name{UMAPcoords}
\alias{UMAPcoords}
\title{UMAP coordinates}
\usage{
UMAPcoords(PiccoloList, Out = T)
}
\arguments{
\item{PiccoloList}{A list object. Piccolo list object obtained after applying the \link[Piccolo]{Normalize} and the \link[Piccolo]{ComputePC} functions}

\item{Out}{A logical variable. Specifies whether to return an output file (.csv) with the normalized values (if set to T), or not (if set to F). Default is T}
}
\value{
A data frame containing the coordinates of the cells in the first 2 UMAP dimensions.
}
\description{
This function will run UMAP for the matrix with the principal components
}
\examples{
\dontrun{
PiccList <- UMAPcoords(PiccoloList = PiccList,
Out = T)
}
}
